+++ classes/book.rb
@@ -1,5 +1,3 @@
-# frozen_string_literal: true
-
 # Book class
 class Book
   attr_accessor :title, :author
