+++ classes/rental.rb
@@ -1,5 +1,3 @@
-# frozen_string_literal: true
-
 # Rental class
 class Rental
   attr_accessor :date
