+++ data/books.json
@@ -1,1 +1,1 @@
-[{"title":"Sizwe Bansi is Dead","author":"Bintu"},{"title":"The chronicles of riddick","author":"Riddika"}]
+[{"title":"Sizwe Bansi is Dead","author":"Bintu"},{"title":"The chronicles of riddick","author":"Riddika"},{"title":"Title","author":"Author"}]
