+++ data/people.json
@@ -1,1 +1,1 @@
-[{"id":4,"age":"12","name":"Flora","classroom":"#<Classroom:0x000002c2394b3b58>","parent_permission":true},{"id":43,"specialization":"MMA Fighter","name":"Ben","age":"13"},{"id":99,"specialization":"MMA Fighter","name":"32","age":"Ron"}]
+[{"id":4,"age":"12","name":"Flora","classroom":"#<Classroom:0x000002c2394b3b58>","parent_permission":true},{"id":43,"specialization":"MMA Fighter","name":"Ben","age":"13"},{"id":99,"specialization":"MMA Fighter","name":"32","age":"Ron"},{"id":25,"age":"32","name":"CodePanther","classroom":"#<Classroom:0x0000561704e39308>","parent_permission":true}]
