+++ data/rentals.json
@@ -1,1 +1,1 @@
-[{"date":"2/2/2","book":{"title":"Sizwe Bansi is Dead","author":"Bintu"},"person":{"id":43,"age":"Ben","name":"13","specialization":"MMA Fighter"}},{"date":"6/6/6","book":{"title":"The chronicles of riddick","author":"Riddika"},"person":{"id":4,"age":"12","name":"Flora","parent_permission":true}},{"date":"8/8/8","book":{"title":"Sizwe Bansi is Dead","author":"Bintu"},"person":{"id":99,"age":"Ron","name":"32","specialization":"MMA Fighter"}}]
+[{"date":"2/2/2","book":{"title":"Sizwe Bansi is Dead","author":"Bintu"},"person":{"id":43,"age":"Ben","name":"13","specialization":"MMA Fighter"}},{"date":"6/6/6","book":{"title":"The chronicles of riddick","author":"Riddika"},"person":{"id":4,"age":"12","name":"Flora","parent_permission":true}},{"date":"8/8/8","book":{"title":"Sizwe Bansi is Dead","author":"Bintu"},"person":{"id":99,"age":"Ron","name":"32","specialization":"MMA Fighter"}},{"date":"20/1/2020","book":{"title":"Title","author":"Author"},"person":{"id":25,"age":"32","name":"CodePanther","parent_permission":true}}]
